experiment-d4rl/requirements.txt
absl-py==1.4.0
accelerate==0.22.0
aiofiles==22.1.0
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.18.0
anyio==3.5.0
appdirs==1.4.4
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
asttokens==2.0.5
async-timeout==4.0.3
attrs==22.1.0
babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
blessed==1.20.0
bottleneck==1.3.5
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.6
cloudpickle==2.2.1
comm==0.1.2
contextlib2==21.6.0
contourpy==1.0.5
cryptography==41.0.2
cycler==0.11.0
cython==0.29.36
# d4rl==1.1
datasets==2.14.4
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.7
dm-control==1.0.14
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
dotmap==1.3.30
entrypoints==0.4
executing==0.8.3
fasteners==0.18
fastjsonschema==2.16.2
filelock==3.12.2
fonttools==4.25.0
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gitdb==4.0.10
gitpython==3.1.32
glfw==2.6.2
gpustat==1.1
gym-notices==0.0.8
gym==0.23.1
h5py==3.9.0
huggingface-hub==0.16.4
idna==3.4
imageio-ffmpeg==0.4.8
imageio==2.31.1
importlib-metadata==6.0.0
importlib-resources==5.2.0
ipykernel==6.25.0
ipython-genutils==0.2.0
ipython==8.12.2
ipywidgets==8.0.4
jedi==0.18.1
jinja2==3.1.2
joblib==1.2.0
json5==0.9.6
jsonschema==4.17.3
jupyter-client==7.4.9
jupyter-console==6.6.3
jupyter-core==5.3.0
jupyter-events==0.6.3
jupyter-server-fileid==0.9.0
jupyter-server-ydoc==0.8.0
jupyter-server==1.23.4
jupyter-ydoc==0.2.4
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.22.0
jupyterlab-widgets==3.0.5
jupyterlab==3.6.3
kiwisolver==1.4.4
kmeans-pytorch==0.3
labmaze==1.0.6
loralib==0.1.1
lxml==4.9.3
markupsafe==2.1.1
matplotlib-inline==0.1.6
matplotlib==3.7.1
mistune==0.8.4
# mjrl==1.0.0
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
ml-collections==0.1.1
moviepy==1.0.3
mujoco-py==2.1.2.14
mujoco==2.3.7
multidict==6.0.4
multiprocess==0.70.15
munkres==1.1.4
mypy-extensions==1.0.0
nbclassic==0.5.5
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.7.0
nest-asyncio==1.5.6
notebook-shim==0.2.2
notebook==6.5.4
numexpr==2.8.4
numpy==1.24.3
nvidia-ml-py==12.535.77
opencv-python==4.8.0.76
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.4.0
pip==23.2.1
pkgutil-resolve-name==1.3.10
platformdirs==2.5.2
ply==3.11
pooch==1.4.0
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==4.24.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pybullet==3.2.5
pycparser==2.21
pygments==2.15.1
pynvml==11.5.0
pyopengl==3.1.7
pyopenssl==23.2.0
pyparsing==3.0.9
pyqt5-sip==12.11.0
pyrallis==0.3.1
pyrsistent==0.18.0
pysocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7
pyyaml==6.0
pyzmq==23.2.0
qtconsole==5.4.2
qtpy==2.2.0
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
send2trash==1.8.0
sentencepiece==0.1.99
sentry-sdk==1.29.2
setproctitle==1.3.2
setuptools==68.0.0
sip==6.6.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.4
stack-data==0.2.0
termcolor==2.3.0
terminado==0.17.1
testpath==0.6.0
threadpoolctl==2.2.0
timm==0.9.5
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch-semiring-einsum==1.1.0
# torch==1.13.1+cu117
# torchaudio==0.13.1+cu117
# torchvision==0.14.1+cu117
tornado==6.3.2
tqdm==4.65.0
traitlets==5.7.1
transformers==4.31.0
typing-extensions==4.7.1
typing-inspect==0.9.0
urllib3==1.26.16
wandb==0.14.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.38.4
widgetsnbextension==4.0.5
xxhash==3.3.0
y-py==0.5.9
yarl==1.9.2
ypy-websocket==0.8.2
zipp==3.11.0
